Restore a compiling workspace after the vyre-libs domain split
The split of vyre-libs into per-domain crates left the workspace unable to
build --all-targets. Four causes, each fixed at its source.

A published item was reachable at two paths: a root re-export beside the
module that defines it. A caller could import either, so a change to one
path silently diverged from the other. Every item now has one path:
vyre, vyre-lower, vyre-reference, vyre-driver, the three concrete drivers,
vyre-runtime, vyre-foundation, vyre-spec and vyre-libs-builder keep the
path their callers already used, and the redundant one is gone.

A feature that gates code importing another crate did not declare the
dependency. vyre-libs now forwards visual and graph-dispatch into
vyre-libs-graph, which declares vyre-libs-visual as the optional
dependency those features need.

A cooperative shared-memory contraction returned a shape mismatch instead
of a kernel. The tiled path now emits one, masking ragged edge tiles so
dimensions that are not multiples of the tile match the reference across
the whole output buffer rather than up to the first partial tile.

Stale declarations and dead code went with them: an orphaned toposort
edge-list module whose callers had already moved to the canonical witness
in vyre-reference, an obsolete security test target, and an all-ends
regex validator pair with no caller.

Two tests asserted the wrong thing and now assert the contract. The
program descriptor expected an empty entry to report no top-level nodes,
but Program::wrapped installs a root region, so it reports one; the same
case now also proves the read-only buffer stays out of the read-write
byte bound, which the previous fixture could not detect. The edge-tile
parity cases compared a prefix of the output and now compare all of it.

+++ vyre-libs-nn/src/nn/linear/layer/linear_4bit/quantized_spec.rs
@@ -199,7 +199,7 @@ mod tests {
                         );
                         assert_eq!(
                             output.output_byte_range(),
-                            Some(0..(out_dim as usize * core::mem::size_of::<f32>())),
+                            Some(0..(u64::from(out_dim) * core::mem::size_of::<f32>() as u64)),
                             "Fix: grouped INT4 output byte range must trim padded launch storage to the logical tensor."
                         );
                         accepted += 1;
